+++ mozilla_django_oidc/views.py
@@ -5,7 +5,6 @@ except ImportError:
     # Python < 3
     from urllib import urlencode
 
-import django
 from django.core.exceptions import SuspiciousOperation
 try:
     from django.urls import reverse
@@ -35,16 +34,20 @@ class OIDCAuthenticationCallbackView(View):
 
     http_method_names = ['get']
 
+    @staticmethod
+    def get_settings(attr, *args):
+        return import_from_settings(attr, *args)
+
     @property
     def failure_url(self):
-        return import_from_settings('LOGIN_REDIRECT_URL_FAILURE', '/')
+        return self.get_settings('LOGIN_REDIRECT_URL_FAILURE', '/')
 
     @property
     def success_url(self):
         # Pull the next url from the session or settings--we don't need to
         # sanitize here because it should already have been sanitized.
         next_url = self.request.session.get('oidc_login_next', None)
-        return next_url or import_from_settings('LOGIN_REDIRECT_URL', '/')
+        return next_url or self.get_settings('LOGIN_REDIRECT_URL', '/')
 
     def login_failure(self):
         return HttpResponseRedirect(self.failure_url)
@@ -54,7 +57,7 @@ class OIDCAuthenticationCallbackView(View):
 
         # Figure out when this id_token will expire. This is ignored unless you're
         # using the RenewIDToken middleware.
-        expiration_interval = import_from_settings('OIDC_RENEW_ID_TOKEN_EXPIRY_SECONDS', 60 * 15)
+        expiration_interval = self.get_settings('OIDC_RENEW_ID_TOKEN_EXPIRY_SECONDS', 60 * 15)
         self.request.session['oidc_id_token_expiration'] = time.time() + expiration_interval
 
         return HttpResponseRedirect(self.success_url)
@@ -112,15 +115,13 @@ def get_next_url(request, redirect_field_name):
     if next_url:
         kwargs = {
             'url': next_url,
-            'require_https': request.is_secure()
+            'require_https': import_from_settings(
+                'OIDC_REDIRECT_REQUIRE_HTTPS', request.is_secure())
         }
 
-        # NOTE(robhudson): Django 2.1 changes `host` to `allowed_hosts`.
-        host = request.get_host()
-        if django.VERSION >= (2, 1):
-            kwargs['allowed_hosts'] = host
-        else:
-            kwargs['host'] = host
+        hosts = list(import_from_settings('OIDC_REDIRECT_ALLOWED_HOSTS', []))
+        hosts.append(request.get_host())
+        kwargs['allowed_hosts'] = hosts
 
         is_safe = is_safe_url(**kwargs)
         if is_safe:
@@ -136,19 +137,23 @@ class OIDCAuthenticationRequestView(View):
     def __init__(self, *args, **kwargs):
         super(OIDCAuthenticationRequestView, self).__init__(*args, **kwargs)
 
-        self.OIDC_OP_AUTH_ENDPOINT = import_from_settings('OIDC_OP_AUTHORIZATION_ENDPOINT')
-        self.OIDC_RP_CLIENT_ID = import_from_settings('OIDC_RP_CLIENT_ID')
+        self.OIDC_OP_AUTH_ENDPOINT = self.get_settings('OIDC_OP_AUTHORIZATION_ENDPOINT')
+        self.OIDC_RP_CLIENT_ID = self.get_settings('OIDC_RP_CLIENT_ID')
+
+    @staticmethod
+    def get_settings(attr, *args):
+        return import_from_settings(attr, *args)
 
     def get(self, request):
         """OIDC client authentication initialization HTTP endpoint"""
-        state = get_random_string(import_from_settings('OIDC_STATE_SIZE', 32))
-        redirect_field_name = import_from_settings('OIDC_REDIRECT_FIELD_NAME', 'next')
-        reverse_url = import_from_settings('OIDC_AUTHENTICATION_CALLBACK_URL',
-                                           'oidc_authentication_callback')
+        state = get_random_string(self.get_settings('OIDC_STATE_SIZE', 32))
+        redirect_field_name = self.get_settings('OIDC_REDIRECT_FIELD_NAME', 'next')
+        reverse_url = self.get_settings('OIDC_AUTHENTICATION_CALLBACK_URL',
+                                        'oidc_authentication_callback')
 
         params = {
             'response_type': 'code',
-            'scope': import_from_settings('OIDC_RP_SCOPES', 'openid email'),
+            'scope': self.get_settings('OIDC_RP_SCOPES', 'openid email'),
             'client_id': self.OIDC_RP_CLIENT_ID,
             'redirect_uri': absolutify(
                 request,
@@ -159,8 +164,8 @@ class OIDCAuthenticationRequestView(View):
 
         params.update(self.get_extra_params(request))
 
-        if import_from_settings('OIDC_USE_NONCE', True):
-            nonce = get_random_string(import_from_settings('OIDC_NONCE_SIZE', 32))
+        if self.get_settings('OIDC_USE_NONCE', True):
+            nonce = get_random_string(self.get_settings('OIDC_NONCE_SIZE', 32))
             params.update({
                 'nonce': nonce
             })
@@ -174,7 +179,7 @@ class OIDCAuthenticationRequestView(View):
         return HttpResponseRedirect(redirect_url)
 
     def get_extra_params(self, request):
-        return import_from_settings('OIDC_AUTH_REQUEST_EXTRA_PARAMS', {})
+        return self.get_settings('OIDC_AUTH_REQUEST_EXTRA_PARAMS', {})
 
 
 class OIDCLogoutView(View):
@@ -182,10 +187,14 @@ class OIDCLogoutView(View):
 
     http_method_names = ['get', 'post']
 
+    @staticmethod
+    def get_settings(attr, *args):
+        return import_from_settings(attr, *args)
+
     @property
     def redirect_url(self):
         """Return the logout url defined in settings."""
-        return import_from_settings('LOGOUT_REDIRECT_URL', '/')
+        return self.get_settings('LOGOUT_REDIRECT_URL', '/')
 
     def post(self, request):
         """Log out the user."""
@@ -194,7 +203,7 @@ class OIDCLogoutView(View):
         if is_authenticated(request.user):
             # Check if a method exists to build the URL to log out the user
             # from the OP.
-            logout_from_op = import_from_settings('OIDC_OP_LOGOUT_URL_METHOD', '')
+            logout_from_op = self.get_settings('OIDC_OP_LOGOUT_URL_METHOD', '')
             if logout_from_op:
                 logout_url = import_string(logout_from_op)(request)
             else: