+++ HISTORY.rst
@@ -3,6 +3,30 @@
 History
 -------
 
+1.2.2 (2019-04-18)
+*******************
+
+* Add Mozilla code of conduct
+* Allow overriding OIDC settings per class
+
+1.2.1 (2019-01-22)
+++++++++++++++++++
+
+* Make `verify_claims` compatible with custom scope configuration.
+
+1.2.0 (2019-01-09)
++++++++++++++++++++
+
+* Improve travis automation for PyPI releases
+* Allow basic auth for OIDC token endpoint requests
+  Thanks `@anttipalola`_
+* Replace phantomjs with firefox headless for e2e testing
+* Add default email verification claim check
+  Thanks `@kerrermanisNL`_
+* Remove compatibility code for unsupported Django versions
+* Add settings to control redirect behavior
+  Thanks `@chrisbrantley`_
+
 1.1.2 (2018-08-24)
 ++++++++++++++++++